+++ CHANGELOG.md
@@ -8,6 +8,9 @@ If you like this project and find it useful, please consider giving it a star on
 
 - [shelly]: Refactor MdnsScanner and MdnsScanner jest test.
 - [shelly]: Added support for battery powered devices.
+- [shelly]: Added component Battery.
+- [shelly]: Added support for shellybutton1.
+
 
 ## [0.7.5] - 2024-07-28